move out message_router
