+++ mpdelta_common/mpdelta_message_router/src/handler.rs
@@ -1,12 +1,12 @@
-use crate::message_router::handler::async_function::{AsyncFunctionHandler, AsyncFunctionHandlerSingle, MessageReceiver};
-use crate::message_router::handler::empty::{EmptyHandler, EmptyHandlerBuilder};
-use crate::message_router::handler::filter_map::FilterMapBuilder;
-use crate::message_router::handler::function::FunctionHandler;
-use crate::message_router::handler::map::MapBuilder;
-use crate::message_router::handler::multiple::MultipleBuilder;
-use crate::message_router::handler::then::ThenBuilder;
-use crate::message_router::static_cow::StaticCow;
-use crate::message_router::MessageHandler;
+use crate::handler::async_function::{AsyncFunctionHandler, AsyncFunctionHandlerSingle, MessageReceiver};
+use crate::handler::empty::{EmptyHandler, EmptyHandlerBuilder};
+use crate::handler::filter_map::FilterMapBuilder;
+use crate::handler::function::FunctionHandler;
+use crate::handler::map::MapBuilder;
+use crate::handler::multiple::MultipleBuilder;
+use crate::handler::then::ThenBuilder;
+use crate::static_cow::StaticCow;
+use crate::MessageHandler;
 use filter::FilterBuilder;
 use mpdelta_async_runtime::AsyncRuntime;
 use std::future::IntoFuture;
