+++ mpdelta_common/mpdelta_message_router/src/handler/empty.rs
@@ -1,6 +1,6 @@
-use crate::message_router::handler::{IntoMessageHandler, MessageHandlerBuilder};
-use crate::message_router::static_cow::StaticCow;
-use crate::message_router::MessageHandler;
+use crate::handler::{IntoMessageHandler, MessageHandlerBuilder};
+use crate::static_cow::StaticCow;
+use crate::MessageHandler;
 use mpdelta_async_runtime::AsyncRuntime;
 use std::marker::PhantomData;
 
@@ -9,7 +9,7 @@ pub struct EmptyHandler;
 pub struct EmptyHandlerBuilder<Runtime>(PhantomData<Runtime>);
 
 impl<Runtime> EmptyHandlerBuilder<Runtime> {
-    pub(in crate::message_router) fn new() -> EmptyHandlerBuilder<Runtime> {
+    pub(crate) fn new() -> EmptyHandlerBuilder<Runtime> {
         EmptyHandlerBuilder(PhantomData)
     }
 }
