+++ mpdelta_common/mpdelta_message_router/src/handler/map.rs
@@ -1,5 +1,5 @@
-use crate::message_router::handler::{IntoMessageHandler, MessageHandler, MessageHandlerBuilder};
-use crate::message_router::static_cow::{Owned, StaticCow};
+use crate::handler::{IntoMessageHandler, MessageHandler, MessageHandlerBuilder};
+use crate::static_cow::{Owned, StaticCow};
 use mpdelta_async_runtime::AsyncRuntime;
 use std::marker::PhantomData;
