+++ mpdelta_common/mpdelta_message_router/src/static_cow.rs
