+++ mpdelta_gui/Cargo.toml
@@ -12,6 +12,7 @@ futures = { workspace = true }
 glam = { workspace = true }
 mpdelta_async_runtime = { workspace = true }
 mpdelta_core = { workspace = true }
+mpdelta_message_router = { workspace = true }
 mpdelta_multimedia = { workspace = true }
 once_cell = { workspace = true }
 ordered-float = { workspace = true }
